[:integtest:pg] Rename PG routine test classes
